Move testes de PaginaInicialView para arquivo próprio #63
